show message when started with -b flag
Show a message when initialization has finished and the user can start the tests
manually when running with the -b flag.

We replace the usage of 'set -x' with a DEBUG trap that prints the executed
commands except echo to stderr, otherwise the echo commands would already be
echoed by 'set -x'.